+++ src/main/java/roidrole/thaumicinfo/ThaumicInformation.java
@@ -16,7 +16,7 @@ import roidrole.thaumicinfo.visualores.PacketHandler;
     name = Tags.MOD_NAME,
     version = Tags.VERSION,
     dependencies =
-        "required-after:jei@[1.12.2-4.15.0.275,);" +
+        "after:jei@[1.12.2-4.15.0.275,);" +
         "required-after:thaumcraft@[6.1.BETA20,);" +
         "required-after:thaumicapi;" +
         "required-after:mixinbooter;" +